[checkpoint_engine] split shard-export helpers: DTensor view stays in fsdp, sparse gather/diff return to delta_sync
Follow-up to the relocation: moving sharded.py wholesale into
verl/workers/engine/fsdp left the checkpoint engine importing from a trainer
backend package -- the inverted dependency the relocation was meant to remove.
Split it instead:

- workers/engine/fsdp/sharded_delta.py keeps only local_shard_view (reads the
  DTensor spec; genuinely the FSDP backend's half of the shard-export contract)
- the framework-agnostic pieces (bytewise shard diff + the sparse gather_v
  collectives, used by any sharded delta sender) return to
  checkpoint_engine/delta_sync/sparse_gather.py
