+++ tests/checkpoint_engine/test_sharded_delta.py
@@ -24,7 +24,8 @@ from __future__ import annotations
 import pytest
 import torch
 
-from verl.workers.engine.fsdp.sharded_delta import local_shard_view, shard_delta_indices
+from verl.checkpoint_engine.delta_sync.sparse_gather import shard_delta_indices
+from verl.workers.engine.fsdp.sharded_delta import local_shard_view
 
 
 @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32])
